pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "email-validator>=2.2.0",
    "flask-cors>=5.0.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "matplotlib>=3.10.1",
    "nltk>=3.9.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.40",
    "tldextract>=5.2.0",
    "trafilatura>=2.0.0",
    "werkzeug>=3.1.3",
    "xgboost>=3.0.0",
]
